training/lidar_lmscnet_style/requirements.txt
torch>=2.1
numpy>=1.24
PyYAML>=6.0
Pillow>=10.0
tqdm>=4.66
tensorboard>=2.14
